docs: update API reference and clarify print request parameters
- Revised the API documentation for the `POST /api/cmd/print` and `POST /api/cmd/print/bulk` endpoints to specify required parameters: `xPosition` and `yPosition`, and the optional `scale`.
- Enhanced descriptions of print settings, emphasizing that only specific fields may be sent in requests, while others must be derived from the server profile.
- Updated error handling documentation to reflect changes in validation rules for `scale`, ensuring it must be greater than 0.
- Improved clarity in the technical documentation regarding print request fields and their expected types.

+++ plotter_signature/domain/contracts.py
@@ -97,7 +97,7 @@ class PrintRequest:
     height: str = "297mm"
     x_position: str = "50mm"
     y_position: str = "50mm"
-    scale: int = 1
+    scale: float = 1.0
     rotation: int = 0
     invert_x: bool = False
     invert_y: bool = True
@@ -111,7 +111,7 @@ class PrintRequest:
             height=str(data.get("height", data.get("Height", "297mm"))),
             x_position=str(data.get("xPosition", data.get("XPosition", "50mm"))),
             y_position=str(data.get("yPosition", data.get("YPosition", "50mm"))),
-            scale=int(data.get("scale", data.get("Scale", 1))),
+            scale=float(data.get("scale", data.get("Scale", 1))),
             rotation=int(data.get("rotation", data.get("Rotation", 0))),
             invert_x=parse_bool(data.get("invertX", data.get("InvertX")), default=False),
             invert_y=parse_bool(data.get("invertY", data.get("InvertY")), default=True),
